Datetime changes and working get().
Using dateutil for datetime parsing.  The serializer is too
open at this point so it serializes too much though.

Add the notion of data_getter and meta_getter to a Service that
are used to flatten an API response into a dict/list container that
acts like the data portion of the response with any meta data fields
set as attributes on the response object.