Consistency fix for async/sync query parsing.
When doing a request who's value is a sequence the standard
way of encoding the url is to repeat the key for each item in
the sequence, a la. ?key=a&key=b.  The 'requests' library does
this automatically, but we need to use the 'doseq' (do sequence)
option on urlencode to replicate the same behavior in the async
adapter which doesn't come with query encoding by default.