Test case updates and rigid trailing_slash behavior
Make the Service(trailing_slash=True) more rigid in that it will always
append the slash regardless of the request arguments.  Previously it
would skip the slash if the path was not provided which I think was an
internal implementation detail due to the path not being of the right
type to append a string to in this case.  I avoid the internal semantics
with the new Service.do impl so it's simple enough to always add the
slash.

Removed unusable test cases which depended on python3.

Added test cases for slash collapasing.

Added test cases for HeaderAuth collapasing.

Updated async.HeaderAuth to support multi-header arguments like the sync
version.